chore: move components into root level directory
